% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProcess.R
\name{annotateGeneName}
\alias{annotateGeneName}
\title{annotate the reName4formatDat with gene info}
\usage{
annotateGeneName(rmNameformatDat, datGRCh)
}
\arguments{
\item{rmNameformatDat}{processed TwoSampleMR::format_data()}

\item{datGRCh}{reference data set with annotation info}
}
\description{
annotate the reName4formatDat with gene info
}
