% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{area}
\alias{area}
\title{Calculate the area under a time-series curve}
\usage{
area(t, x)
}
\arguments{
\item{t}{the time}

\item{x}{the value at each time}
}
\description{
Calculate the area under a time-series curve
}
