% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProcess.R
\name{checkSNPsOverlap}
\alias{checkSNPsOverlap}
\title{calculate number of the overlapped SNPs between two traits given a pval threshold}
\usage{
checkSNPsOverlap(dat1, dat2, pval)
}
\arguments{
\item{dat1}{GWAS summary statistics from \code{TwoSampleMR::format_data()} for trait 1}

\item{dat2}{GWAS summary statistics from \code{TwoSampleMR::format_data()} for trait 2}

\item{pval}{pvalue threshod for selecting the overlapped SNPs}
}
\description{
calculate number of the overlapped SNPs between two traits given a pval threshold
}
