% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProcess.R
\name{dat4GRAPPLE}
\alias{dat4GRAPPLE}
\title{prepare data for implementing MR-GRAPPLE}
\usage{
dat4GRAPPLE(dat)
}
\arguments{
\item{dat}{formated exposue- or outcome-GWAS using TwoSampleMR::format_data()}
}
\description{
prepare data for implementing MR-GRAPPLE
}
