% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProcess.R
\name{datMerge}
\alias{datMerge}
\title{merge the data set based on SNP}
\usage{
datMerge(datX, datY)
}
\arguments{
\item{datX}{data frame X. Usually, from \code{rmName4formatDat()}}

\item{datY}{data frame Y. Usually, from \code{rmName4FormatDat()}}
}
\description{
merge the data set based on SNP
}
