% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProcess.R
\name{getAnnotatedSNPs}
\alias{getAnnotatedSNPs}
\title{Get annotated SNPs info}
\usage{
getAnnotatedSNPs(dat, datGRCh, label)
}
\arguments{
\item{dat}{data frame obtained from \code{TwoSampleMR::format_data()} for exposure GWAS}

\item{datGRCh}{reference panel for gene annotation in either GRCh37 or GRCh38}

\item{label}{label the trait or phenotype info}
}
\description{
Get annotated SNPs info
}
