% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProcess.R
\name{getLDmatrix}
\alias{getLDmatrix}
\title{Get LDmatrix for given SNPs concerning 1KG reference}
\usage{
getLDmatrix(dat, ldRef, pval)
}
\arguments{
\item{dat}{data frame from \code{TwoSampleMR::format_data()}}

\item{ldRef}{location of 1KG reference panel}

\item{pval}{pvalue threshold for selecting SNPs}
}
\description{
Get LDmatrix for given SNPs concerning 1KG reference
}
