% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProcess.R
\name{getOverlapSNPs}
\alias{getOverlapSNPs}
\title{get the overlapped SNPs between two traits for colocalization analyses}
\usage{
getOverlapSNPs(dat1, trait1, dat2, trait2, ldRef, pval)
}
\arguments{
\item{dat1}{GWAS summary statistics from \code{TwoSampleMR::format_data()} for trait1}

\item{trait1}{trait name for dat1}

\item{dat2}{GWAS summary statistics from \code{TwoSAmpleMR::format_data()} for trait2}

\item{trait2}{trait name for dat2}

\item{ldRef}{1KG reference panel in plink format}

\item{pval}{pvalue threshold for selecting SNPs with \code{pvalue<pval}}
}
\description{
get the overlapped SNPs between two traits for colocalization analyses
}
