% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{loadSPSS}
\alias{loadSPSS}
\title{Import data files with SPSS format}
\usage{
loadSPSS(name, path)
}
\arguments{
\item{name}{The name of the SPSS data files}

\item{path}{The path of the SPSS data files}
}
\description{
Import data files with SPSS format
}
