% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt.R
\name{plt.stack.region}
\alias{plt.stack.region}
\title{plot the regional association plot for exp-, eQTL- and out-GWAS}
\usage{
plt.stack.region(
  geneInfo,
  datOut,
  outName,
  datQTL,
  qtlName,
  datExp,
  expName,
  ldRef
)
}
\arguments{
\item{geneInfo}{dataset contains the gene info that is generally obtained from \code{datGRCh37} or \code{datGRCh38}.}

\item{datOut}{out-GWAS summary statistics obtained from \code{TwoSampleMR::format_data()}}

\item{outName}{name of the outcome trait}

\item{datQTL}{QTL-GWAS summary statistics obtained from \code{TwoSampleMR::format_data()} or with the same data format}

\item{qtlName}{name of the eQTL trait}

\item{datExp}{exp-GWAS summary statistics obtained from \code{TwoSampleMR::format_data()}}

\item{expName}{name of the exposure trait}

\item{ldRef}{1KG reference panel in plink format}
}
\description{
plot the regional association plot for exp-, eQTL- and out-GWAS
}
