% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt.R
\name{pltManhattan}
\alias{pltManhattan}
\title{plot Manhattan plot}
\usage{
pltManhattan(
  datManhattan,
  leadingSNPs,
  filename,
  type = c("circle", "manhattan")
)
}
\arguments{
\item{datManhattan}{data frame with the first three column being \code{c("SNP","Chromosome","Position")} with the following columns being \code{pval} for traits of interest}

\item{leadingSNPs}{data frame with at least \code{c("SNP","uniprot_gn_symbol")} info}

\item{filename}{location and name of the figure}

\item{type}{the type of the Manhattan plot, with \code{type=c("circle","manhattan")}}
}
\description{
plot Manhattan plot
}
