% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGWASCatalog.R
\name{readdeCODEProteomicsGRCh38}
\alias{readdeCODEProteomicsGRCh38}
\title{read summary statistics from deCODE using Olink Explore 1536 Panel}
\usage{
readdeCODEProteomicsGRCh38(filename, type)
}
\arguments{
\item{filename}{location and file name of a specific data set}

\item{type}{type of the data set, \code{type=c("exposure","outcome")}}
}
\description{
read summary statistics from deCODE using Olink Explore 1536 Panel
}
