% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGWASCatalog.R
\name{readHarmonizedGWASCatalogGRCh37}
\alias{readHarmonizedGWASCatalogGRCh37}
\title{read harmonized summary statistics from GWAS Catalog}
\usage{
readHarmonizedGWASCatalogGRCh37(filename, type)
}
\arguments{
\item{filename}{location and file name of a specific data set}

\item{type}{type of the data set, \code{type=c("exposure","outcome")}}
}
\description{
read harmonized summary statistics from GWAS Catalog
}
