% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProcess.R
\name{rmName4formatDat}
\alias{rmName4formatDat}
\title{rename TwoSampleMR::format_data() with snps selection at \code{pval}}
\usage{
rmName4formatDat(dat, label, pval)
}
\arguments{
\item{dat}{data frame obtained from TwoSampleMR::format_data()}

\item{label}{label the data frame}

\item{pval}{p-value threshod for selecting significant snps}
}
\description{
rename TwoSampleMR::format_data() with snps selection at \code{pval}
}
