% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProcess.R
\name{rnDat}
\alias{rnDat}
\title{remove ".exposure" and ".outcome" from a data frame}
\usage{
rnDat(dat)
}
\arguments{
\item{dat}{GWAS summary statistics from \code{TwoSampleMR::format_data()}}
}
\description{
remove ".exposure" and ".outcome" from a data frame
}
\keyword{internal}
