% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProcess.R
\name{run_SuSiE}
\alias{run_SuSiE}
\title{Implement the fine-mapping analysis using SuSiE method}
\usage{
run_SuSiE(dat, ldRef, pval, nSampleSize, threads, withAlleles = TRUE)
}
\arguments{
\item{dat}{data frame from \code{TwoSampleMR::format_data()} for exposure GWAS}

\item{ldRef}{location of 1KG reference panel}

\item{pval}{pvalue threshold for selecting SNPs}

\item{nSampleSize}{sample size of the exposure GWAS}

\item{threads}{number of threads used for plink}

\item{withAlleles}{indicator of whether LD matrix includes effect and other alleles}
}
\description{
Implement the fine-mapping analysis using SuSiE method
}
