% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMR.R
\name{runCisMR}
\alias{runCisMR}
\title{run drug target (cis-) MR analyses with annotated data}
\usage{
runCisMR(datExp, datMR, FDR.level = 0.01)
}
\arguments{
\item{datExp}{data frame for annotated exposure GWAS using \code{annotateGeneName()}}

\item{datMR}{data frame from \code{TwoSampleMR::harmonised_data()}}

\item{FDR.level}{threshold for controlling false positive rate using q-value}
}
\description{
run drug target (cis-) MR analyses with annotated data
}
