% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMR.R
\name{runDebiasedMR}
\alias{runDebiasedMR}
\title{run de-biased MR using TwoSampleMR pkg}
\usage{
runDebiasedMR(datMR)
}
\arguments{
\item{datMR}{the harmonized data for MR analysis using TwoSampleMR::harmonise()}
}
\description{
run de-biased MR using TwoSampleMR pkg
}
