% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMR.R
\name{runGeneMendelianRandomization}
\alias{runGeneMendelianRandomization}
\title{run Gene-based MR analyses with correlated IVs}
\usage{
runGeneMendelianRandomization(datMR, ldRef, pval = 1e-04, corr = TRUE)
}
\arguments{
\item{datMR}{data frame from \code{TwoSampleMR::harmonised_data()}}

\item{ldRef}{1KG reference panel for getting LD matrices}

\item{pval}{threshold for selecting IVs at p-value}

\item{corr}{\code{corr=c(TRUE,FALSE)} indicator of inclusion of correlated IVs or not}
}
\description{
run Gene-based MR analyses with correlated IVs
}
