% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMR.R
\name{runGRAPPLE}
\alias{runGRAPPLE}
\title{run three-sample MR using MR-GRAPPLE}
\usage{
runGRAPPLE(
  selGWAS.file,
  expGWAS.file,
  outGWAS.file,
  pval,
  ldRef,
  clump_r2 = 0.001,
  outPath
)
}
\arguments{
\item{selGWAS.file}{dataset with selection GWAS summary statistics}

\item{expGWAS.file}{dataset with exposure GWAS summary statistics}

\item{outGWAS.file}{dataset with outcome GWAS summary statistics}

\item{pval}{threshold of p-value for SNPs selection}

\item{ldRef}{location of ld reference panel}

\item{clump_r2}{the r2 used for clumping}

\item{outPath}{the path for saving figures generated from MR-GRAPPLE}
}
\description{
run three-sample MR using MR-GRAPPLE
}
