% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMR.R
\name{runMendelianRandomization}
\alias{runMendelianRandomization}
\title{run MR using \pkg{MendelianRandomization}}
\usage{
runMendelianRandomization(datMRInput)
}
\arguments{
\item{datMRInput}{data frame from \code{MendelianRandomization::mr_input()}}
}
\description{
run MR using \pkg{MendelianRandomization}
}
