% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMR.R
\name{runMRClust}
\alias{runMRClust}
\title{run MR-Clust analysis}
\usage{
runMRClust(datMR)
}
\arguments{
\item{datMR}{the harmonized data for MR analysis using TwoSampleMR::harmonise()}
}
\description{
run MR-Clust analysis
}
