% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMR.R
\name{runMRClustBest}
\alias{runMRClustBest}
\title{run MR analyses based on MR-Clust}
\usage{
runMRClustBest(datMR, datMRClust.best)
}
\arguments{
\item{datMR}{data frame from TwoSampleMR::harmonised_data()}

\item{datMRClust.best}{data frame obtained from MR-Clust with best selection}
}
\description{
run MR analyses based on MR-Clust
}
