% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMR.R
\name{runMRIVW}
\alias{runMRIVW}
\title{run MR-IVW using TwoSampleMR pkg}
\usage{
runMRIVW(datMR)
}
\arguments{
\item{datMR}{the harmonized data for MR analysis using TwoSampleMR::harmonise()}
}
\description{
run MR-IVW using TwoSampleMR pkg
}
