% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMR.R
\name{runMRRAPS}
\alias{runMRRAPS}
\title{run MR-RAPS using TwoSampleMR pkg}
\usage{
runMRRAPS(datMR)
}
\arguments{
\item{datMR}{the harmonized data for MR analysis using TwoSampleMR::harmonise()}
}
\description{
run MR-RAPS using TwoSampleMR pkg
}
