% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMR.R
\name{runSuSiERMR}
\alias{runSuSiERMR}
\title{run MR analysis based on SuSiER}
\usage{
runSuSiERMR(datGeneMR, ldRef, clump_r2 = 0.2)
}
\arguments{
\item{datGeneMR}{data frame with gene-level data from \code{TwoSampleMR::harmonised_data()}}

\item{ldRef}{1KG reference panel for getting LD matrices}

\item{clump_r2}{correlation in LD. By default, \code{clump_r2=0.2}}
}
\description{
run MR analysis based on SuSiER
}
