% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProcess.R
\name{dat4UKBNealeLab}
\alias{dat4UKBNealeLab}
\title{data processing for UKB-Neale Labs GWAS summary statistics}
\usage{
dat4UKBNealeLab(UKBGWAS.file, datUKBSNP, type = "outcome", binaryTrait = TRUE)
}
\arguments{
\item{UKBGWAS.file}{path of the GWAS summary statistics}

\item{datUKBSNP}{processed variants info obtained from variants.tsv.gz with \code{c("variant","rsid","consequence","consequence_category")}}

\item{type}{the type of exposure- or outcome- GWAS processed using \code{TwoSampleMR::format_data()}. By default, \code{type="outcome"}}

\item{binaryTrait}{an indicator of a binary trait or not. By default, \code{binaryTrait=TRUE}.}
}
\description{
data processing for UKB-Neale Labs GWAS summary statistics
}
