% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProcess.R
\name{plink_clump}
\alias{plink_clump}
\title{clumping the data}
\usage{
plink_clump(
  dat,
  ldRef,
  clump_kb = 10000,
  clump_r2 = 0.001,
  clump_p1 = 1,
  clump_p2 = 1,
  threads = 36,
  tempdir = "temp"
)
}
\arguments{
\item{dat}{data frame with at least two columns names \code{c("SNP","pval.exposure")}}

\item{ldRef}{1KG reference panel with .bed format}

\item{clump_kb}{clumping window. By default, \code{clump_kb=10000}}

\item{clump_r2}{clumping r2 cutoff. By default, \code{clump_r2=0.001}}

\item{clump_p1}{p-value threshold for the index SNP. By default, \code{clump_p1=1}}

\item{clump_p2}{p-value threshold for the secondary SNP. By default, \code{clump_p2=1}}

\item{threads}{number of threads used for plink}

\item{tempdir}{temporal direct for storing plink data set}
}
\description{
clumping the data
}
