% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataProcess.R
\name{run_clump}
\alias{run_clump}
\title{run ld clump locally}
\usage{
run_clump(formatDatExp, ldRef)
}
\arguments{
\item{formatDatExp}{data frame from \code{TwoSampleMR::format_data()} for exposure GWAS}

\item{ldRef}{1KG refernce panel in plink format}
}
\description{
run ld clump locally
}
